Update IT-glue-BitLocker-Documentation.ps1
As the recovery password doesn't always exist in the first array element, converting to string will allow it to print out the key regardless if its in the first or second array element. This does add a space character to the output though at the beginning or end depending on the array element it is in.